requirements.txt
datasets==3.6.0
flashgeotext==0.5.5
gensim==4.3.0
gensim==3.8.3
joblib==1.3.2
joblib==1.2.0
joblib==1.0.1
joblib==1.4.2
keybert==0.8.5
matplotlib==3.5.3
matplotlib==3.5.1
matplotlib==3.0.3
matplotlib==3.3.4
matplotlib==3.7.5
multi_rake==0.0.2
networkx==2.3
networkx==3.4.2
networkx==3.0
networkx==3.3
networkx==2.4
networkx==2.5
nltk==3.6.1
nltk==3.9.1
numpy==1.21.6
numpy==1.23.2
numpy==1.22.2
numpy==1.18.5
numpy==1.19.5
numpy==1.15.1
numpy==1.20.1
numpy==1.21.5
numpy==1.24.4
pandas==1.2.1
pandas==1.5.2
pandas==1.2.4
pandas==1.3.5
pandas==2.0.3
Pillow==9.1.0
Pillow==9.2.0
Pillow==9.0.1
Pillow==8.4.0
Pillow==9.3.0
Pillow==8.2.0
Pillow==11.3.0
pmdarima==2.0.4
pyecharts==1.9.1
pyecharts==2.0.6
PyMySQL==1.0.2
rake_nltk==1.0.6
scikit_learn==1.0.2
scikit_learn==1.2.0
scikit_learn==1.1.2
scikit_learn==0.24.1
scikit_learn==1.3.2
scipy==1.7.3
scipy==1.9.1
scipy==1.6.2
seaborn==0.11.1
seaborn==0.10.1
spacy==3.7.5
statsmodels==0.13.2
statsmodels==0.12.2
statsmodels==0.14.1
torch==1.13.1
torch==2.6.0
torch==2.3.1
torch==2.4.1
tornado==6.2
tornado==6.1
tqdm==4.67.0
tqdm==4.64.1
tqdm==4.66.5
tqdm==4.67.1
tqdm==4.66.2
tqdm==4.59.0
transformers==4.49.0
transformers==4.45.2
wordcloud==1.9.4
XlsxWriter==3.2.0
XlsxWriter==1.3.8
